+++ inject_test.go
@@ -1,6 +1,7 @@
 package service
 
 import (
+	"context"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -16,7 +17,7 @@ func TestInject(t *testing.T) {
 	container := New()
 	container.Set("value", &T1{42})
 	obj := &T2{}
-	err := Inject(container, obj)
+	err := Inject(context.Background(), container, obj)
 	require.Nil(t, err)
 	assert.Equal(t, 42, obj.Value.val)
 }
@@ -30,7 +31,7 @@ func TestInjectNonPointer(t *testing.T) {
 	container := New()
 	container.Set("value", T1{42})
 	obj := &T2{}
-	err := Inject(container, obj)
+	err := Inject(context.Background(), container, obj)
 	require.Nil(t, err)
 	assert.Equal(t, 42, obj.Value.val)
 }
@@ -45,7 +46,7 @@ func TestInjectAnonymous(t *testing.T) {
 	container := New()
 	container.Set("value", &T1{42})
 	obj := &T3{&T2{}}
-	err := Inject(container, obj)
+	err := Inject(context.Background(), container, obj)
 	require.Nil(t, err)
 	assert.Equal(t, 42, obj.Value.val)
 }
@@ -60,7 +61,7 @@ func TestInjectAnonymousZeroValue(t *testing.T) {
 	container := New()
 	container.Set("value", &T1{42})
 	obj := &T3{} // not &T3{&T2{}}
-	err := Inject(container, obj)
+	err := Inject(context.Background(), container, obj)
 	require.Nil(t, err)
 	assert.Equal(t, 42, obj.Value.val)
 }
@@ -75,7 +76,7 @@ func TestInjectAnonymousNonPointer(t *testing.T) {
 	container := New()
 	container.Set("value", &T1{42})
 	obj := &T3{}
-	err := Inject(container, obj)
+	err := Inject(context.Background(), container, obj)
 	require.Nil(t, err)
 	assert.Equal(t, 42, obj.Value.val)
 }
@@ -87,7 +88,7 @@ func TestInjectAnonymousZeroValueNoServiceTags(t *testing.T) {
 	container := New()
 	container.Set("value", &T1{42})
 	obj := &T2{}
-	err := Inject(container, obj)
+	err := Inject(context.Background(), container, obj)
 	require.Nil(t, err)
 	assert.Nil(t, obj.T1)
 }
@@ -102,7 +103,7 @@ func TestInjectAnonymousUnexported(t *testing.T) {
 	container := New()
 	container.Set("value", &T1{42})
 	obj := &T3{&t2{}}
-	err := Inject(container, obj)
+	err := Inject(context.Background(), container, obj)
 	require.Nil(t, err)
 	assert.Nil(t, obj.t2.Value)
 }
@@ -110,7 +111,7 @@ func TestInjectAnonymousUnexported(t *testing.T) {
 func TestInjectNonStruct(t *testing.T) {
 	container := New()
 	obj := func() error { return nil }
-	err := Inject(container, obj)
+	err := Inject(context.Background(), container, obj)
 	require.Nil(t, err)
 }
 
@@ -122,7 +123,7 @@ func TestInjectMissingService(t *testing.T) {
 
 	container := New()
 	obj := &T2{}
-	err := Inject(container, obj)
+	err := Inject(context.Background(), container, obj)
 	assert.EqualError(t, err, `no service registered to key "value"`)
 }
 
@@ -136,7 +137,7 @@ func TestInjectBadType(t *testing.T) {
 	container := New()
 	container.Set("value", &T3{3.14})
 	obj := &T2{}
-	err := Inject(container, obj)
+	err := Inject(context.Background(), container, obj)
 	assert.EqualError(t, err, "field 'Value' cannot be assigned a value of type *service.T3")
 }
 
@@ -149,7 +150,7 @@ func TestInjectNil(t *testing.T) {
 	container := New()
 	container.Set("value", nil)
 	obj := &T2{}
-	err := Inject(container, obj)
+	err := Inject(context.Background(), container, obj)
 	assert.EqualError(t, err, "field 'Value' cannot be assigned a value of type nil")
 }
 
@@ -161,12 +162,12 @@ func TestInjectOptional(t *testing.T) {
 
 	container := New()
 	obj := &T2{}
-	err := Inject(container, obj)
+	err := Inject(context.Background(), container, obj)
 	require.Nil(t, err)
 	require.Nil(t, obj.Value)
 
 	container.Set("value", &T1{42})
-	err = Inject(container, obj)
+	err = Inject(context.Background(), container, obj)
 	require.Nil(t, err)
 	assert.Equal(t, 42, obj.Value.val)
 }
@@ -179,7 +180,7 @@ func TestInjectBadOptional(t *testing.T) {
 
 	container := New()
 	obj := &T2{}
-	err := Inject(container, obj)
+	err := Inject(context.Background(), container, obj)
 	assert.EqualError(t, err, "field 'Value' has an invalid optional tag")
 }
 
@@ -191,6 +192,6 @@ func TestContainerUnsettableFields(t *testing.T) {
 
 	container := New()
 	container.Set("value", &T1{42})
-	err := Inject(container, &T2{})
+	err := Inject(context.Background(), container, &T2{})
 	assert.EqualError(t, err, "field 'value' can not be set - it may be unexported")
 }