+++ post_inject.go
@@ -1,7 +1,9 @@
 package service
 
+import "context"
+
 // PostInject is a marker interface for injectable objects which should
 // perform some action after injection of services.
 type PostInject interface {
-	PostInject() error
+	PostInject(ctx context.Context) error
 }